pyproject.toml
[project]
name = "telugu-voice-agent"
version = "0.1.0"
dependencies = [
    "langchain-groq",
    "groq",
    "edge-tts",
    "sounddevice",
    "scipy",
    "langgraph",
    "python-dotenv",
    "langchain-community"
]